Checked the package for class
Comments are labels with the "Max - " string Added some imports, changed some boolean statements for clarity.
Changed the function 'FormattedQ2xts' with regular expressions to show another way of doing it.
Edited the testing_script.R file to provide some additional things ways of doing it.
Overall, its really good.  You lay out your methods and process very clearly and it is easy to follow/ review. The dygraph function is a nice function and I'll have to check it out. Hope this helps a bit! - Max